#117 separate gender domain model
